+++ CHANGELOG.md
@@ -14,10 +14,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add tab key handler to indent/unindent list items
 - Add eslint-plugin-deprecation package to check for use of deprecated code.
 - Disable undo and redo button when there's no more steps
+- Add support for ordered list styles
+- Added the ability to pass optional attributes to the `toggleList` command.
 - Addition of spacing between paragraphs/tables and other elements.
 
 ### Changed
 - Increase left-padding of lists
+- BREAKING: Rename Toolbar::Icon to Toolbar::Button
 - remove overflow: hidden on editor container div
 - Replace drone by woodpecker
 - Modernize toolbar dropdown